Revert "Dynamize box price in tokens and date"
