+++ src/config/constants/Guildpads/TankWarsZone2.ts
@@ -45,8 +45,6 @@ const guildpad = {
       }
     ]
   },
-  projectTokenEquivalent: '1425 WBOND',
-  asOfPriceInProjectToken: 1645362000000,
   epochEndDate: 1645362000000,
 }
