+++ src/views/GuildPad/Pad/Cards/index.tsx
@@ -19,7 +19,6 @@ import TokenLogo from 'components/Launchpad/Logo'
 import { PostBody, NavOption, SaleContainer, SaleRow } from '../../../../components/Launchpad/styled'
 import Boxcard from '../BoxCard'
 import '../../../../css/styleFX.css'
-import Content from '../../../../components/Launchpad/Content'
 
 
 const GCard = styled(SCard)<{ src?: string }>`
@@ -310,6 +309,106 @@ const CardHeader: React.FC<{ status: string; background?: string; guildpad: Guil
   </Header>
 )
 
+// PROJECT CONTENT
+const Content: React.FC<{ guildpad: GuildpadConfig,  userDataLoaded: boolean, rarity?: string }> = ({
+  guildpad,
+  userDataLoaded,
+  rarity = '1',
+}) => {
+  const theme = useContext(ThemeContext)
+  const [active, setActive] = useState(1)
+
+  const renderDescription = () => {
+    const description = guildpad.description !== '' ? guildpad.description : 'No description'
+
+    return (
+      <Text color="textSubtle" margin="10px 0px" style={{ lineHeight: '2em' }}>
+        {description}
+      </Text>
+    )
+  }
+
+  const renderSale = () => {
+    const price = guildpad.boxInfo[rarity].price !== '0' ? guildpad.boxInfo[rarity].price : 'TBA'
+    const ratio = guildpad.inoDetails.ratio && guildpad.inoDetails.ratio
+    const boxes = guildpad.inoDetails.boxes ?? 'TBA'
+    const start = guildpad.date.start ?? 'TBA'
+    const end = guildpad.date.end ?? 'TBA'
+    const distribution = guildpad.distribution
+    const initMarketCap = 'TBA'
+    const initTokenCirc = 'TBA'
+
+    return (
+      <SaleContainer justifyContent="space-between">
+        <Flex flexDirection="column">
+          <SaleRow justifyContent="space-between">
+            <Text color="textSubtle">Sale Price</Text>
+            <div style={{textAlign: 'right'}}>
+              <Text>{price !== 'TBA' ? `${price} ${guildpad.buyingCoin.symbol} (1425 WBOND)` : price }</Text>
+              { price !== 'TBA' && <Text fontSize='12px'>(<em>as of Feb 18, 2022 8PM GMT+8</em>)</Text> }
+            </div>
+          </SaleRow>
+          {ratio && (
+            <SaleRow justifyContent="space-between">
+              <Text color="textSubtle">Ratio</Text>
+              <Text>{guildpad.inoDetails.ratio}</Text>
+            </SaleRow>
+          )}
+          <SaleRow justifyContent="space-between">
+            <Text color="textSubtle">Sale Start Time</Text>
+            <Text>{start}</Text>
+          </SaleRow>
+          <SaleRow justifyContent="space-between">
+            <Text color="textSubtle">Sale End Time</Text>
+            <Text>{end}</Text>
+          </SaleRow>
+        </Flex>
+        <Flex flexDirection="column">
+          {guildpad.type === TYPE.INO && (
+            <SaleRow justifyContent="space-between">
+              <Text color="textSubtle">Boxes for Sale</Text>
+              <Text>{boxes}</Text>
+            </SaleRow>
+          )}
+          <SaleRow justifyContent="space-between">
+            <Text color="textSubtle">{guildpad.type === TYPE.INO ? 'NFT' : 'TOKEN'} Distribution</Text>
+            <Text>{distribution}</Text>
+          </SaleRow>
+          {guildpad.type === TYPE.IDO && (
+            <>
+              <SaleRow justifyContent="space-between">
+                <Text color="textSubtle">Initial Market Cap</Text>
+                <Text>{initMarketCap}</Text>
+              </SaleRow>
+              <SaleRow justifyContent="space-between">
+                <Text color="textSubtle">Initial Token Circulation</Text>
+                <Text>{initTokenCirc}</Text>
+              </SaleRow>
+            </>
+          )}
+        </Flex>
+      </SaleContainer>
+    )
+  }
+
+  return (
+    <PostBody style={{ zIndex: 1, backgroundColor: `${theme.colors.MGG_container}` }}>
+      <Flex
+        alignItems="center"
+        margin="10px 0px 20px 0px"
+        style={{ borderBottom: `0.5px solid ${theme.colors.primary}`, width: '100%' }}
+      >
+        <NavOption onClick={() => setActive(1)} activeIndex={active === 1}>
+          Project Description
+        </NavOption>
+        <NavOption onClick={() => setActive(2)} activeIndex={active === 2}>
+          {guildpad.type === TYPE.INO? 'NFT' : 'TOKEN'} Sale
+        </NavOption>
+      </Flex>
+      {active === 1 ? renderDescription() : active === 2 && renderSale()}
+    </PostBody>
+  )
+}
 
 const Card: React.FC<{ guildpad: GuildpadConfig, userDataLoaded: boolean }> = ({ guildpad, userDataLoaded }) => {
   const { account } = useWeb3React()
@@ -330,7 +429,7 @@ const Card: React.FC<{ guildpad: GuildpadConfig, userDataLoaded: boolean }> = ({
           <Boxcard imgProps={{ src: 'Chest3.png', size: '15rem' }} guildpad={guildpad} userDataLoaded={userDataLoaded} />
         </ContainerBoxCard>
         <ContainerProjDesc>
-          <Content guildpad={guildpad} />
+          <Content guildpad={guildpad} userDataLoaded={userDataLoaded} />
         </ContainerProjDesc>
       </Contain>
     </GCard>
